requirements.txt
crawl4ai
streamlit
playwright 
google-generativeai
schedule
